% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLP.R
\name{blp_regression}
\alias{blp_regression}
\title{BLP Estimation Strategies}
\usage{
blp_regression(
  y,
  D,
  cates,
  strategy = "wr",
  denoise = "none",
  pscore = NULL,
  mu = NULL,
  mu0 = NULL,
  mu1 = NULL,
  scores = NULL
)
}
\arguments{
\item{y}{Observed outcomes.}

\item{D}{Treatment indicator.}

\item{cates}{Estimated CATEs. CATEs must be estimated with different observations than those in \code{y} and \code{D}.}

\item{strategy}{String indicating the desired identification strategy. One of "wr", "ht", "aipw".}

\item{denoise}{String denoting whether to include optional constructed covariates to the regression. One of "none", "cddf1", "cddf2", "mck1", "mck2", "mck3". Useful to reduce the variance of the estimation.}

\item{pscore}{Estimated propensity scores. They must be estimated with different observations than those in \code{y} and \code{D}. Necessary only under certain combinations of \code{strategy} and \code{denoise}.}

\item{mu}{Estimated regression function. It must be estimated with different observations than those in \code{y} and \code{D}. Necessary only under certain combinations of \code{strategy} and \code{denoise}.}

\item{mu0}{Estimated regression function for control units. It must be estimated with different observations than those in \code{y} and \code{D}. Necessary only under certain combinations of \code{strategy} and \code{denoise}.}

\item{mu1}{Estimated regression function for treated. It must be estimated with different observations than those in \code{y} and \code{D}. Necessary only under certain combinations of \code{strategy} and \code{denoise}.}

\item{scores}{Estimated doubly-robust scores. They must be estimated with different observations than those in \code{y} and \code{D}. Necessary only if \code{strategy} is set to \code{"aipw"}.}
}
\value{
Return.
}
\description{
Estimates the best linear predictor of the actual CATEs using the estimated CATEs.
}
\details{
Currently, the routine supports only single-splits.
}
\examples{
## Generate data.
set.seed(1986)

n <- 3000
k <- 3

}
\seealso{
Other functions
}
\author{
Riccardo Di Francesco
}
