% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{compute_aipw_scores}
\alias{compute_aipw_scores}
\title{AIPW Pseudo-Outcomes}
\usage{
compute_aipw_scores(Y, D, pscore, mu0, mu1)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment indicator vector, binary 0/1 with 1 for treated.}

\item{pscore}{Propensity score estimates of P(D = 1 | X).}

\item{mu0}{Conditional mean estimates of E[Y | D = 0, X].}

\item{mu1}{Conditional mean estimates of E[Y | D = 1, X].}
}
\value{
Numeric vector of AIPW pseudo-outcomes.
}
\description{
Computes AIPW pseudo-outcomes from nuisance function estimates.
}
\seealso{
\code{\link{estimate_nuisances}}, \code{\link{valiCATE}}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
