% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{compute_weights}
\alias{compute_weights}
\title{Compute Weight Values}
\usage{
compute_weights(tau_hat, spec)
}
\arguments{
\item{tau_hat}{Numeric vector of CATE predictions.}

\item{spec}{Weight specification list from \code{\link{get_weight_spec}}.}
}
\value{
A list with the following elements:\cr

\code{omega}: Numeric vector of weight values omega_hat(X_i).\cr
\code{v}: Numeric vector of centering values v_hat(X_i).\cr
\code{nabla}: Numeric vector of gradient values nabla_v_h(X_i).
}
\description{
Given CATE predictions and a weight specification, computes the weight values omega_hat(X_i), the values v_hat(X_i),
and the gradient values nabla_v_h(X_i) for each observation.
}
\references{
\itemize{
  \item Di Francesco, R., & Knaus, M. C. (2025). Validating ML Predictions of Heterogeneous Treatment Effects via CWATE.
}
}
\seealso{
\code{\link{get_weight_spec}}, \code{\link{valiCATE}}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
