% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cates-evaluation.R
\name{evalue_cates}
\alias{evalue_cates}
\title{CATEs Evaluation}
\usage{
evalue_cates(
  y,
  D,
  X,
  cates,
  is_train,
  pscore = NULL,
  strategy = "wr",
  denoise = "none"
)
}
\arguments{
\item{y}{Observed outcomes.}

\item{D}{Treatment indicator.}

\item{X}{Covariate matrix (no intercept).}

\item{cates}{Estimated CATEs. CATEs must be estimated using only the training sample.}

\item{is_train}{Logical vector denoting which observations belong to the training sample.}

\item{pscore}{Estimated propensity scores. They must be estimated using only the training sample. If not provided by the user, they are estimated internally via an honest regression forest.}

\item{strategy}{String indicating the desired identification strategy. One of "wr", "ht", "aipw".}

\item{denoise}{String denoting whether to include optional constructed covariates to the regressions. One of "none", "cddf1", "cddf2", "mck1", "mck2", "mck3". Useful to reduce the variance of the estimation.}
}
\value{
Return.
}
\description{
Evaluates the quality of CATEs estimates by implementing several Generic Machine Learning methodologies.
}
\details{
Currently, the routine supports only single-splits.\cr
Regression functions are estimated via honest \code{\link[grf]{regression_forest}}s.
}
\examples{
## Generate data.
set.seed(1986)

n <- 3000
k <- 3

}
\seealso{
Other functions
}
\author{
Riccardo Di Francesco
}
