% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_weight_spec}
\alias{get_weight_spec}
\title{Weight Function Specifications}
\usage{
get_weight_spec(weight_name)
}
\arguments{
\item{weight_name}{Character string, one of \code{"AUTOC"}, \code{"AUC-HVL"}, \code{"BLP"}, \code{"QINI"}.}
}
\value{
A list with the following elements:\cr

\code{name}: Character name of the weight.\cr
\code{h}: Function h(t, v) defining omega(x) = h(tau_hat(x), v(x)).\cr
\code{nabla_v_h}: Function computing the partial derivative of h with respect to v.\cr
\code{cdf_based}: Logical. If \code{TRUE}, v(x) = F(tau_hat(x)) and m_x(t) = I(t <= tau_hat(x)).
  If \code{FALSE} (BLP), v(x) = E[tau_hat] and m_x(t) = t.
}
\description{
Returns a list describing the weight function for a given weight name.
}
\details{
The four weight functions are defined as follows:\cr

AUTOC: h(t, v) = -log(1 - v) - 1, with nabla_v_h = 1 / (1 - v).\cr
AUC-HVL: h(t, v) = log(v / (1 - v)), with nabla_v_h = 1 / (v * (1 - v)).\cr
BLP: h(t, v) = t - v, with nabla_v_h = -1.\cr
QINI: h(t, v) = v - 1/2, with nabla_v_h = 1.\cr

For CDF-based weights (AUTOC, AUC-HVL, QINI), v is computed using a safe empirical CDF that avoids 0 and 1 values.
For BLP, v is the sample mean of the CATE predictions.
}
\references{
\itemize{
  \item Di Francesco, R., & Knaus, M. C. (2025). Validating ML Predictions of Heterogeneous Treatment Effects via CWATE.
}
}
\seealso{
\code{\link{compute_weights}}, \code{\link{valiCATE}}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
