% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{print.valiCATE}
\alias{print.valiCATE}
\title{Print Method for valiCATE Objects}
\usage{
\method{print}{valiCATE}(x, ...)
}
\arguments{
\item{x}{An \code{\link{valiCATE}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Prints an \code{\link{valiCATE}} object.
}
\description{
Prints an \code{\link{valiCATE}} object.
}
\references{
\itemize{
  \item Di Francesco, R., & Knaus, M. C. (2025). Validating ML Predictions of Heterogeneous Treatment Effects via CWATE.
}
}
\seealso{
\code{\link{valiCATE}}
}
\author{
Riccardo Di Francesco
}
