% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.valiCATE}
\alias{summary.valiCATE}
\title{Summary Method for valiCATE Objects}
\usage{
\method{summary}{valiCATE}(object, latex = FALSE, ...)
}
\arguments{
\item{object}{An \code{\link{valiCATE}} object.}

\item{latex}{Logical, whether to print LATEX code for a table. If \code{TRUE}, a ready-to-compile LATEX table with CWATE and
NCWATE point estimates, confidence intervals, and p-values is printed to the console.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Summarizes an \code{\link{valiCATE}} object.
}
\description{
Summarizes an \code{\link{valiCATE}} object.
}
\details{
Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}.
}
\examples{
\donttest{## Generate data.
set.seed(1986)

n <- 1000
k <- 2

X <- matrix(rnorm(n * k), ncol = k)
colnames(X) <- paste0("x", seq_len(k))
D <- rbinom(n, size = 1, prob = 0.5)
mu0 <- 0.5 * X[, 1]
mu1 <- 0.5 * X[, 1] + X[, 2]
Y <- mu0 + D * (mu1 - mu0) + rnorm(n, sd = 0.5)

## Split into training and validation samples.
train_idx <- sample(1:n, n / 2)
val_idx <- setdiff(1:n, train_idx)

## Estimate CATEs on the training sample, predict on the validation sample.
library(grf)
cf <- causal_forest(X[train_idx, ], Y[train_idx], D[train_idx])
cates <- predict(cf, X[val_idx, ])$predictions

## Validate using the validation sample.
result <- valiCATE(Y[val_idx], D[val_idx], X[val_idx, ],
                   cates = list("causal_forest" = cates))

## Summarize.
summary(result)
summary(result, latex = TRUE)}

}
\references{
\itemize{
  \item Di Francesco, R., & Knaus, M. C. (2025). Validating ML Predictions of Heterogeneous Treatment Effects via CWATE.
}
}
\seealso{
\code{\link{valiCATE}}
}
\author{
Riccardo Di Francesco
}
