% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary_console}
\alias{summary_console}
\title{Console Summary}
\usage{
summary_console(object)
}
\arguments{
\item{object}{An \code{\link{valiCATE}} object.}
}
\value{
Console output.
}
\description{
Displays CWATE and NCWATE results in the console.
}
\keyword{internal}
