% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary_latex}
\alias{summary_latex}
\title{LaTeX Summary}
\usage{
summary_latex(object)
}
\arguments{
\item{object}{An \code{\link{valiCATE}} object.}
}
\value{
LATEX code printed to console.
}
\description{
Produces LATEX table code for CWATE and NCWATE results.
}
\details{
Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}.
}
\keyword{internal}
