% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATES.R
\name{test_parametric_gates}
\alias{test_parametric_gates}
\title{Testing for Effect Heterogeneity using the estimated GATES}
\usage{
test_parametric_gates(model)
}
\arguments{
\item{model}{Estimated parametric model, as one of those returned by \code{\link{gates_estimation}}.}
}
\value{
A list with the p-values for the three hypotheses tested.
}
\description{
Tests for effect heterogeneity using the estimated parametric model for the sorted group average treatment effects (GATES).
}
\details{
\code{model} must consist of a \code{lm_robust} object where the coefficients identifying the GATES must be called \code{"group1"}, \code{"group2"}, and so on.\cr

Three distinct hypotheses of effect heterogeneity are tested: whether all GATES are equal to each other, whether the largest and the smallest GATES are different from each other, 
and whether the differences in the GATES across all pairs of groups are zero. For the last test, we adjust p-values to account for multiple hypotheses testing using Holm's procedure 
and report the median of the adjusted p-values.
}
\seealso{
\code{\link{gates_estimation}}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
