% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLP.R
\name{blp_estimation}
\alias{blp_estimation}
\title{BLP Estimation}
\usage{
blp_estimation(
  Y,
  D,
  cates,
  pscore,
  mu,
  mu0,
  mu1,
  scores,
  strategies,
  denoising
)
}
\arguments{
\item{Y}{Observed outcomes.}

\item{D}{Treatment indicator.}

\item{cates}{CATE predictions. Must be produced by a model estimated using different observations than those in \code{Y} and \code{D}.}

\item{pscore}{Propensity scores predictions. Must be produced by a model estimated using different observations than those in \code{Y} and \code{D} (unless the propensity score is known, in which case we provide the true values).}

\item{mu}{Conditional mean predictions. Must be produced by a model estimated using different observations than those in \code{Y} and \code{D}}

\item{mu0}{Control units' conditional mean predictions. Must be produced by a model estimated using different observations than those in \code{Y} and \code{D}}

\item{mu1}{Treated units' conditional mean predictions. Must be produced by a model estimated using different observations than those in \code{Y} and \code{D}}

\item{scores}{Estimated doubly-robust scores. Must be estimated via K-fold cross-fitting using the same observations as in \code{Y} and \code{D}.}

\item{strategies}{Character vector controlling the identification strategies to implement for BLP and GATES. Admitted values are \code{"WR"} (weighted residuals), \code{"HT"} (Horwitz-Thompson), and \code{"AIPW"} (augmented inverse-probability weighting).}

\item{denoising}{Character vector controlling if and which additional covariates to include in the regressions to reduce the variance of the estimation. Admitted values are \code{"none"}, \code{"cddf1"}, \code{"cddf2"}, \code{"mck1"}, \code{"mck2"}, and \code{"mck3"}.}
}
\value{
A list of fitted models as \code{\link[estimatr]{lm_robust}} objects.
}
\description{
Estimates the best linear predictor (BLP) of the actual CATEs using the estimated CATEs.
}
\details{
To estimate the BLP of the actual CATEs using the estimated CATEs, the user must provide observations on the outcomes and the treatment status of units in 
the validation sample, as well as their estimated cates and nuisance functions. Be careful, as these estimates must be obtained using only observations from the training sample (see the example section below).
Additionally, the user must provide doubly-robust scores estimated in the validation sample using K-fold cross fitting.\cr

The BLP is estimated using three different strategies, all involving fitting suitable linear models. For each of these strategis, different model specifications are considered that differ in additional and
optional covariates that can be included in the regressions to reduce the estimation variance. Check the online \href{https://riccardo-df.github.io/valiCATE/articles/valiCATE-short-tutorial.html}{short tutorial}
and \href{https://riccardo-df.github.io/valiCATE/articles/denoising.html}{denoising vignette} for details.\cr

Standard errors are estimated using the Eicker-Huber-White estimator.
}
\seealso{
\code{\link{gates_estimation}}, \code{\link{toc_estimation}}, \code{\link{rate_estimation}}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
