% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATES.R
\name{blp_regression}
\alias{blp_regression}
\title{GATES Estimation}
\usage{
blp_regression(y, D, cates, pscore, mu, mu0, mu1, scores)
}
\arguments{
\item{y}{Observed outcomes.}

\item{D}{Treatment indicator.}

\item{cates}{Estimated CATEs. CATEs must be estimated with different observations than those in \code{y} and \code{D}.}

\item{pscore}{Propensity scores. If unknown, they must be estimated using different observations than those in \code{y} and \code{D}.}

\item{mu}{Estimated regression function. It must be estimated with different observations than those in \code{y} and \code{D}.}

\item{mu0}{Estimated regression function for control units. It must be estimated with different observations than those in \code{y} and \code{D}.}

\item{mu1}{Estimated regression function for treated units. It must be estimated with different observations than those in \code{y} and \code{D}.}

\item{scores}{Estimated doubly-robust scores. They must be estimated via K-fold cross-fitting.}
}
\value{
Return.
}
\description{
Estimates the group average treatment effects, with the groups formed by cutting the distribution of the estimated CATEs into K quantiles.
}
\details{
Currently, the routine supports only single-splits.\cr
We rearrange the estimate GATES to obey the monotonicity property.
}
\examples{
## Generate data.
set.seed(1986)

n <- 3000
k <- 3

}
\seealso{
Other functions
}
\author{
Riccardo Di Francesco
}
