% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cates-evaluation.R
\name{evalue_cates}
\alias{evalue_cates}
\title{CATEs Evaluation}
\usage{
evalue_cates(y, D, X, cates, is_train, pscore = NULL, verbose = TRUE)
}
\arguments{
\item{y}{Observed outcomes.}

\item{D}{Treatment indicator.}

\item{X}{Covariate matrix (no intercept).}

\item{cates}{Estimated CATEs. CATEs must be estimated using only the training sample.}

\item{is_train}{Logical vector denoting which observations belong to the training sample.}

\item{pscore}{Propensity scores. If unknown, they must be estimated using only the training sample. If not provided by the user, they are estimated internally via an honest \code{\link[grf]{regression_forest}}.}

\item{verbose}{Logical, set to FALSE to prevent the function from printing the progresses.}
}
\value{
Return.
}
\description{
Evaluates the quality of CATEs estimates by implementing several Generic Machine Learning methodologies.
}
\details{
Currently, the routine supports only single-splits.\cr
Regression functions are estimated via honest \code{\link[grf]{regression_forest}}s.\cr
If strategy is set to \code{"aipw"}, doubly-robust scores are estimated internally using the validation sample via 5-fold cross fitting and using honest regression forests
(see the \code{\link[aggTrees]{dr_scores}} function).\cr
If not provided, \code{cates} are estimated internally using a causal forest estimator and the training sample.\cr
If not provided, \code{pscore} are estimated internally using a regression forest.\cr
}
\examples{
## Generate data.
set.seed(1986)

n <- 3000
k <- 3

}
\seealso{
Other functions
}
\author{
Riccardo Di Francesco
}
