% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cates-evaluation.R
\name{evalue_cates}
\alias{evalue_cates}
\title{CATEs Evaluation}
\usage{
evalue_cates(
  Y,
  D,
  X,
  cates,
  is_train,
  pscore = NULL,
  mu = NULL,
  mu0 = NULL,
  mu1 = NULL,
  n_groups = 5,
  beneficial = TRUE,
  n_boot = 200,
  verbose = TRUE
)
}
\arguments{
\item{Y}{Observed outcomes.}

\item{D}{Treatment indicator.}

\item{X}{Covariate matrix (no intercept).}

\item{cates}{Estimated CATEs. Must be estimated using only the training sample.}

\item{is_train}{Logical vector denoting which observations belong to the training sample.}

\item{pscore}{Propensity scores. If unknown, must be estimated using only the training sample.}

\item{mu}{Estimated regression function. Must be estimated using only the training sample.}

\item{mu0}{Estimated regression function for control units. Must be estimated using only the training sample.}

\item{mu1}{Estimated regression function for treated units. Must be estimated using only the training sample.}

\item{n_groups}{Number of groups to be formed for the GATES analysis.}

\item{beneficial}{Logical, whether the treatment is beneficial to units. If \code{TRUE}, units are ranked according to decreasing values of \code{cates} to estimate the RATE, otherwise they are ranked according to increasing values of \code{cates}.}

\item{n_boot}{Number of bootstrap replications to estimate the standard error of the RATE estimate.}

\item{verbose}{Logical, set to FALSE to prevent the function from printing the progresses.}
}
\value{
An \code{evaluCATE} object.
}
\description{
Evaluates the quality of CATEs estimates by estimating the best linear predictor (BLP) of the actual CATEs using the estimated
CATEs, the sorted group average treatment effects (GATES), and the rank-weighted average treatment effect (RATE) induced by
the estimated CATEs.
}
\details{
To estimate BLP, GATES, and RATE, the user must provide observations on the outcomes, the treatment status, and the covariates of units in the whole sample, as well
as their estimated CATEs. Be careful, as the CATEs must be estimated only with part of the sample, which we call the training sample (see the example section below).\cr

To let the function know which observations were used for the CATEs estimation, the user must also provide a logical vector with the \code{TRUE}s denoting observations in the
training sample. This way, \code{\link{evalue_cates}} knows which observations to use to post-process the CATEs estimates.\cr

\code{\link{evalue_cates}} implements a number of strategies to estimate the BLP and the GATES. Most of them involve fitting a suitable linear model. The linear models differ according to the
different identification strategies. Furthermore, for each strategy, there exist various sets of constructed covariates that one can add to reduce the variance of the estimation. \code{\link{evalue_cates}}
fits and returns all these possible models. GATES are also estimated using a nonparametric approach. Check the online
\href{https://riccardo-df.github.io/evaluCATE/articles/evalue-cates-short-tutorial.html}{short tutorial} for details.\cr

Some of the linear models involve covariates that depend on particular nuisance functions, e.g., propensity score and conditional mean of the outcome
(check the online \href{https://riccardo-df.github.io/evaluCATE/articles/denoising.html}{denoising vignette} for details about these covariates). The user can supply estimates of these functions by using the
optional arguments \code{pscore}, \code{mu}, \code{mu0}, and \code{mu1}. Be careful, as these must be obtained using only the training sample. If not provided by the user, these functions are estimated internally
via honest \code{\link[grf]{regression_forest}}s using only the training sample. \cr

For the linear models, standard errors are estimated using the Eicker-Huber-White estimator.\cr

To estimate the BLP and GATES using the AIPW strategy, doubly-robust scores are estimated internally using the validation sample via 5-fold cross fitting and honest regression forests (see the
\code{\link[aggTrees]{dr_scores}} function for details).\cr

The estimated GATES are sorted to enforce monotonicity.\cr

Two different RATEs are estimated: AUTOC and QINI coefficient. Sample-averaging estimators are employed. Standard errors are estimated by the standard deviation of the bootstrap estimates obtained using the half-sample bootstrap.
}
\examples{
## Generate data.
set.seed(1986)

n <- 1000
k <- 2

X <- matrix(rnorm(n * k), ncol = k)
colnames(X) <- paste0("x", seq_len(k))
D <- rbinom(n, size = 1, prob = 0.5)
mu0 <- 0.5 * X[, 1]
mu1 <- 0.5 * X[, 1] + X[, 2]
Y <- mu0 + D * (mu1 - mu0) + rnorm(n)

## Sample split.
train_idx <- sample(c(TRUE, FALSE), length(Y), replace = TRUE)

X_tr <- X[train_idx, ]
X_val <- X[!train_idx, ]

D_tr <- D[train_idx]
D_val <- D[!train_idx]

Y_tr <- Y[train_idx]
Y_val <- Y[!train_idx]

## CATEs estimation.
library(grf)

forest <- causal_forest(X_tr, Y_tr, D_tr) # We use only the training sample.
cates <- predict(forest, X)$predictions # We predict on the whole sample.

## CATEs evaluation. Estimate all nuisances internally. 
pscore <- rep(0.5, length(Y))
evaluation <- evalue_cates(Y, D, X, cates, train_idx, pscore = pscore)

## Generic S3 methods.
summary(evaluation)
summary(evaluation, latex = "BLP")

plot(evaluation, target = "GATES")
plot(evaluation, target = "RATE")

}
\seealso{
Other functions
}
\author{
Riccardo Di Francesco
}
