% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{plot.evaluCATE}
\alias{plot.evaluCATE}
\title{Plot Method for evaluCATE Objects}
\usage{
\method{plot}{evaluCATE}(
  x,
  target = "GATES",
  which_models = names(x$GATES),
  gates_hline = TRUE,
  toc_smoother = 1,
  ...
)
}
\arguments{
\item{x}{An \code{\link{evaluCATE}} object.}

\item{target}{String controlling which plot to display. Must be one of \code{"GATES"}, \code{"TOC"}, or \code{"RATE"}.}

\item{which_models}{Character vector controlling which results to display. Admitted values are those stored in \code{names(x$GATES)}. Ignored if \code{target != "GATES"}.}

\item{gates_hline}{Logical, whether to display an horizontal line at zero in the GATES plot. Ignored if \code{target != "GATES"}.}

\item{toc_smoother}{Integer number, controls the amount of smoothing in the \code{TOC} and \code{RATE} plots. Smoothing is achieved by plotting the TOCs only for every other \code{toc_smoother} unit (in order of treatment benefit). Set to 1 to plot the TOC of all units, to 2 to plot the TOC of every other unit, ecc. Ignored if \code{target == "GATES"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Plots an \code{\link{evaluCATE}} object.
}
\description{
Plots an \code{\link{evaluCATE}} object.
}
\details{
Check the online \href{https://riccardo-df.github.io/evaluCATE/articles/more-on-plotting.html}{plotting vignette} for details.
}
\author{
Riccardo Di Francesco
}
