% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{print.evaluCATE}
\alias{print.evaluCATE}
\title{Print Method for evaluCATE Objects}
\usage{
\method{print}{evaluCATE}(x, target = "BLP", latex = FALSE, which_models = c("all"), ...)
}
\arguments{
\item{x}{An \code{evaluCATE} object.}

\item{target}{String controlling which plot to display. Must be either \code{"BLP"} or \code{"GATES"}.}

\item{latex}{Logical, whether to print LATEX code for a table. Different tables are produced according to the \code{target} argument.}

\item{which_models}{Character vector with the names of the models to report in the LATEX table. Set equal to \code{c("all")} to display all models. Ignored if \code{latex == FALSE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Prints an \code{evaluCATE} object.
}
\description{
Prints an \code{evaluCATE} object.
}
\details{
Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}.
}
\examples{
\donttest{## Generate data.
set.seed(1986)

n <- 1000
k <- 2

X <- matrix(rnorm(n * k), ncol = k)
colnames(X) <- paste0("x", seq_len(k))
D <- rbinom(n, size = 1, prob = 0.5)
mu0 <- 0.5 * X[, 1]
mu1 <- 0.5 * X[, 1] + X[, 2]
Y <- mu0 + D * (mu1 - mu0) + rnorm(n)

## Sample split.
train_idx <- sample(c(TRUE, FALSE), length(Y), replace = TRUE)

X_tr <- X[train_idx, ]
X_val <- X[!train_idx, ]

D_tr <- D[train_idx]
D_val <- D[!train_idx]

Y_tr <- Y[train_idx]
Y_val <- Y[!train_idx]

## CATEs estimation.
library(grf)

forest <- causal_forest(X_tr, Y_tr, D_tr) # We use only the training sample.
cates <- predict(forest, X)$predictions # We predict on the whole sample.

## CATEs evaluation. Estimate all nuisances internally. 
pscore <- rep(0.5, length(Y))
evaluation <- evalu_cates(Y, D, X, cates, train_idx, pscore = pscore)

## Print.
print(evaluation, target = "BLP")
print(evaluation, target = "BLP", latex = TRUE)

print(evaluation, target = "GATES")
print(evaluation, target = "GATES", latex = TRUE)}

}
\author{
Riccardo Di Francesco
}
