% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.evaluCATE}
\alias{summary.evaluCATE}
\title{Summary Method for evaluCATE Objects}
\usage{
\method{summary}{evaluCATE}(
  object,
  target = "BLP",
  latex = FALSE,
  which_models = names(object$GATES[[1]]),
  ...
)
}
\arguments{
\item{object}{An \code{\link{evaluCATE}} object.}

\item{target}{String controlling which parameters we are interested in. Must be one of \code{"BLP"}, \code{"GATES"}, \code{"RATE"}.}

\item{latex}{Logical, whether to print LATEX code for a table. Different tables are produced according to the \code{target} argument.}

\item{which_models}{Character vector controlling which results to display. Admitted values are those stored in \code{names(object$GATES[[1]])}. Ignored if \code{target == "RATE"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Summarizes an \code{\link{evaluCATE}} object.
}
\description{
Summarizes an \code{\link{evaluCATE}} object.
}
\details{
Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}, \code{multirow}.
}
\seealso{
\code{\link{evaluCATE}}
}
\author{
Riccardo Di Francesco
}
