% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.evaluCATE}
\alias{summary.evaluCATE}
\title{Summary Method for evaluCATE Objects}
\usage{
\method{summary}{evaluCATE}(object, latex = "none", ...)
}
\arguments{
\item{object}{An \code{evaluCATE} object.}

\item{latex}{String to denote whether and which table to print in LATEX code. Must be one of \code{"none"}, \code{"BLP"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Summarizes an \code{evaluCATE} object.
}
\description{
Summarizes an \code{evaluCATE} object.
}
\details{
Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}. If
standard errors have been estimated, they are printed in parenthesis below each point estimate.
}
\examples{
## Generate data.
set.seed(1986)

n <- 1000
k <- 2

X <- matrix(rnorm(n * k), ncol = k)
colnames(X) <- paste0("x", seq_len(k))
D <- rbinom(n, size = 1, prob = 0.5)
mu0 <- 0.5 * X[, 1]
mu1 <- 0.5 * X[, 1] + X[, 2]
Y <- mu0 + D * (mu1 - mu0) + rnorm(n)

## Sample split.
train_idx <- sample(c(TRUE, FALSE), length(Y), replace = TRUE)

X_tr <- X[train_idx, ]
X_val <- X[!train_idx, ]

D_tr <- D[train_idx]
D_val <- D[!train_idx]

Y_tr <- Y[train_idx]
Y_val <- Y[!train_idx]

## CATEs estimation.
library(grf)

forest <- causal_forest(X_tr, Y_tr, D_tr) # We use only the training sample.
cates <- predict(forest, X)$predictions # We predict on the whole sample.

## CATEs evaluation. Estimate all nuisances internally. 
pscore <- rep(0.5, length(Y))
evaluation <- evalue_cates(Y, D, X, cates, train_idx, pscore = pscore)

## Summary.
summary(evaluation)
summary(evaluation, latex = "BLP")

}
\seealso{
\code{\link{evalue_cates}}
}
\author{
Riccardo Di Francesco
}
