% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RATE.R
\name{toc_estimation}
\alias{toc_estimation}
\title{TOC Estimation}
\usage{
toc_estimation(cates, scores, beneficial = TRUE)
}
\arguments{
\item{cates}{Estimated CATEs. Must be out-of-sample predictions on the validation sample.}

\item{scores}{Estimated doubly-robust scores. Must be estimated via K-fold cross-fitting using the validation sample.}

\item{beneficial}{Logical, whether the treatment is beneficial to units. If \code{TRUE}, units are ranked according to decreasing values of \code{cates}, otherwise they are ranked according to increasing values of \code{cates}.}
}
\value{
A vector of estimated TOCs sorted according to the increasing or decreasing values of \code{cates}.
}
\description{
Estimates the targeting operator characteristic (TOC) induced by the estimated CATEs.
}
\details{
To estimate the TOC induced by the estimated CATEs, the user must provide the estimated CATEs and doubly-robust scores. Be careful, as the CATEs must be estimated only with 
using the training sample, while the doubly-robust scores must be estimated in the validation sample using K-fold cross fitting (see the example section below).\cr

The TOC is estimated using a sample-averaging estimator. Check the \href{https://riccardo-df.github.io/evaluCATE/articles/evaluCATE-short-tutorial.html}{online vignette} 
for details.
}
\seealso{
\code{\link{blp_estimation}}, \code{\link{gates_estimation}}, \code{\link{rate_estimation}}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
